refactor jest matcher type definition
