+++ src/to-match-dot-json-snapshot.ts
@@ -1,19 +1,6 @@
 import { toMatchSnapshot } from 'jest-snapshot';
 import { jsonStdin } from './lib/dot-adapter';
 
-declare global {
-  namespace jest {
-    // tslint:disable-next-line: interface-name
-    interface Expect {
-      toMatchDotJsonSnapshot<T>(): JestMatchers<T>;
-    }
-    // tslint:disable-next-line: interface-name
-    interface Matchers<R, T> {
-      toMatchDotJsonSnapshot(): R;
-    }
-  }
-}
-
 export function toMatchDotJsonSnapshotMatcher(this: any, dot: string): jest.CustomMatcherResult {
   try {
     return toMatchSnapshot.call(this, jsonStdin(dot));
